Rename util functions
