+++ packages/ui/src/components/Dialogs/KeystoresExitDialog.tsx
@@ -18,11 +18,11 @@ import {
   CustomValidatorGetResponse,
   ValidatorExitExecute,
   shortenPubkey,
-  Network
+  Network,
 } from "@stakingbrain/common";
 import { api } from "../../api";
 import { SlideTransition } from "./Transitions";
-import { getEmoji } from "../../utils/dataUtils";
+import { getEmoji } from "../../utils/data";
 
 export default function KeystoresExitDialog({
   rows,
@@ -37,7 +37,7 @@ export default function KeystoresExitDialog({
   setSelectedRows: (selectedRows: GridSelectionModel) => void;
   open: boolean;
   setOpen: (open: boolean) => void;
-  network: Network
+  network: Network;
 }): JSX.Element {
   const [validatorsExitResponse, setValidatorsExitResponse] =
     useState<ValidatorExitExecute[]>();
@@ -147,7 +147,11 @@ export default function KeystoresExitDialog({
                     id="alert-dialog-description"
                     component={"span"}
                   >
-                    <ExitWarning rows={rows} selectedRows={selectedRows} network={network} />
+                    <ExitWarning
+                      rows={rows}
+                      selectedRows={selectedRows}
+                      network={network}
+                    />
                   </DialogContentText>
 
                   <br />
